\name{revComplement-methods}
\docType{methods}
\alias{revComplement-methods}
\alias{revComplement,MassArrayData-method}
\alias{revComplement,character-method}
\title{ Reverse complement (methods)}
\description{
 Methods for reverse complement
}
\section{Methods}{
\describe{

\item{x = "MassArrayData"}{ Finds reverse complement of a MassArrayData object, a function that consists of altering sequence, strand, fragmentation, and methylation data }

\item{x = "character"}{ Calculates reverse complement from nucleotide sequence as character input }
}}
\author{ Reid F. Thompson (\email{rthompso@aecom.yu.edu}), John M. Greally (\email{jgreally@aecom.yu.edu}) }
\seealso{ \code{\link{revComplement}} }
\keyword{methods}
