\name{rnaDigest}
\alias{rnaDigest}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ RNA digest }
\description{
  Function to perform an in silico RNAse A digest for either the T or C cleavage reactions
}
\usage{
rnaDigest(sequence, type = c("T", "C"))
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sequence}{ Nucleotide sequence input }
  \item{type}{ One of either 'T' or 'C', specifying which cleavage reaction mixture was used }
}
\value{
  Returns a list of \code{MassArrayFragment} objects, each containing information about a given fragment generated by the RNA digest
}
\author{ Reid F. Thompson (\email{rthompso@aecom.yu.edu}), John M. Greally (\email{jgreally@aecom.yu.edu}) }
\examples{
rnaDigest("AAAACCCCCTGCGGAGAGAGGCCGACAAAA", type="T")
}
\seealso{ See Also as \code{\linkS4class{MassArrayFragment}} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ character }
\keyword{ manip }
